feat: move provider around, feature-gated callback helpers
