+++ packages/blitz-net/src/backend/reqwest.rs
@@ -1,5 +1,12 @@
+use std::sync::Arc;
+
+use crate::{ProviderError, USER_AGENT};
+
 use super::{BackendError, RequestBackend, Response};
-use blitz_traits::net::Request;
+use blitz_traits::net::{BoxedHandler, Bytes, NetCallback, NetProvider, Request, SharedCallback};
+use data_url::DataUrl;
+use http::HeaderValue;
+use tokio::{runtime::Handle, sync::mpsc::{unbounded_channel, UnboundedReceiver, UnboundedSender}};
 
 // Compat with reqwest
 impl From<reqwest::Error> for BackendError {
@@ -42,4 +49,47 @@ impl RequestBackend for Backend {
             body,
         })
     }
+}
+
+pub struct Provider<D> {
+  rt: Handle,
+  client: Backend,
+  resource_callback: SharedCallback<D>,
+}
+
+impl<D: 'static> Provider<D> {
+    pub fn new(res_callback: SharedCallback<D>) -> Self {
+        Self {
+            rt: Handle::current(),
+            client: Backend::new(),
+            resource_callback: res_callback,
+        }
+    }
+
+    pub fn shared(res_callback: SharedCallback<D>) -> Arc<dyn NetProvider<Data = D>> {
+        Arc::new(Self::new(res_callback))
+    }
+
+    pub fn is_empty(&self) -> bool {
+        Arc::strong_count(&self.resource_callback) == 1
+    }
+}
+
+impl<D: 'static> NetProvider for Provider<D> {
+  type Data = D;
+
+  fn fetch(&self, doc_id: usize, request: Request, handler: BoxedHandler<D>) {
+      let client = self.client.clone();
+      let callback = Arc::clone(&self.resource_callback);
+      println!("Fetching {}", &request.url);
+      drop(self.rt.spawn(async move {
+          let url = request.url.to_string();
+          let res = Self::fetch_inner(client, doc_id, request, handler, callback).await;
+          if let Err(e) = res {
+              eprintln!("Error fetching {}: {e}", url);
+          } else {
+              println!("Success {}", url);
+          }
+      }));
+  }
 }
